requirements.txt
fastapi[all]
sqlmodel              
psycopg2-binary
pwdlib[argon2]
pyjwt
python-multipart